Se lleva a un módulo aparte la llamada al modelo
